View modification started

+++ database/migrations/2021_08_03_064027_add_mobile_to_users.php
@@ -15,7 +15,7 @@ class AddMobileToUsers extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             //
-            $table->integer('mobile')->after('email');
+            $table->bigInteger('mobile')->after('email');
         });
     }
